+++ CHANGELOG.md
@@ -8,6 +8,9 @@
     - `CopyButton` a ready-to-use copy button with customizable text, icons, and colors to indicate copy state.
     - `CustomCopyButton` a fully customizable component using JavaScript function props for advanced use-cases.
 
+- Added function as prop support for `xAxisProps`, `yAxisProps`, `gridProps`, `rightYAxisProps` (and `zAxisProps` for BubbleChart) in `AreaChart`, `BarChart`, `BubbleChart`, `CompositeChart`, `LineChart`, and `ScatterChart`. #661 by @AnnMarieW
+
+
 # 2.3.0